backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
Pillow==10.4.0
python-dotenv==1.0.1
django-storages==1.14.4
boto3==1.34.144
gunicorn==22.0.0
django-filter==24.2
drf-spectacular==0.27.2
whitenoise==6.7.0
